+++ src/app/app.module.ts
@@ -9,9 +9,9 @@ import { AuthGuard } from '../auth/guard/auth/auth.guard';
 import { PrismaModule } from '../prisma/prisma.module';
 import { MailModule } from '../mail/mail.module';
 import { ProviderModule } from '../provider/provider.module';
-import { ReviewsModule } from 'src/reviews/reviews.module';
-import { ConnectionsModule } from 'src/connections/connections.module';
-import { AppLoggerMiddleware } from 'src/middlewares/logger.middleware';
+import { ReviewsModule } from '../../src/reviews/reviews.module';
+import { ConnectionsModule } from '../../src/connections/connections.module';
+import { AppLoggerMiddleware } from '../../src/middlewares/logger.middleware';
 import { CommonModule } from '../common/common.module';
 
 @Module({
