+++ src/connections/connections.controller.spec.ts
@@ -1,5 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ConnectionsController } from './connections.controller';
+import { ConnectionsService } from './connections.service';
+import { PrismaService } from '../../src/prisma/prisma.service';
 
 describe('ConnectionsController', () => {
   let controller: ConnectionsController;
@@ -7,6 +9,7 @@ describe('ConnectionsController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [ConnectionsController],
+      providers: [ConnectionsService, PrismaService],
     }).compile();
 
     controller = module.get<ConnectionsController>(ConnectionsController);
