+++ src/connections/connections.service.spec.ts
@@ -1,12 +1,13 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ConnectionsService } from './connections.service';
+import { PrismaService } from '../../src/prisma/prisma.service';
 
 describe('ConnectionsService', () => {
   let service: ConnectionsService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [ConnectionsService],
+      providers: [ConnectionsService, PrismaService],
     }).compile();
 
     service = module.get<ConnectionsService>(ConnectionsService);
