+++ src/connections/connections.service.ts
@@ -4,10 +4,10 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { AuthType, User } from '@prisma/client';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { PrismaService } from '../../src/prisma/prisma.service';
 import { ConnectionDto } from './dto/Connection.dto';
 import { ApiTags } from '@nestjs/swagger';
-import { ProfileFetcherDelegator } from 'src/user/profile-fetcher/delegator.profile-fetcher';
+import { ProfileFetcherDelegator } from '../../src/user/profile-fetcher/delegator.profile-fetcher';
 import { v4 } from 'uuid';
 
 @Injectable()
