+++ src/reviews/reviews.controller.spec.ts
@@ -1,5 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ReviewsController } from './reviews.controller';
+import { ReviewsService } from './reviews.service';
+import { PrismaService } from '../../src/prisma/prisma.service';
+import { REDIS_CLIENT } from '../../src/provider/redis.provider';
 
 describe('ReviewsController', () => {
   let controller: ReviewsController;
@@ -7,6 +10,14 @@ describe('ReviewsController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [ReviewsController],
+      providers: [
+        ReviewsService,
+        PrismaService,
+        {
+          provide: REDIS_CLIENT,
+          useValue: {},
+        },
+      ],
     }).compile();
 
     controller = module.get<ReviewsController>(ReviewsController);
