+++ src/reviews/reviews.service.spec.ts
@@ -1,12 +1,21 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ReviewsService } from './reviews.service';
+import { PrismaService } from '../../src/prisma/prisma.service';
+import { REDIS_CLIENT } from '../../src/provider/redis.provider';
 
 describe('ReviewsService', () => {
   let service: ReviewsService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [ReviewsService],
+      providers: [
+        ReviewsService,
+        PrismaService,
+        {
+          provide: REDIS_CLIENT,
+          useValue: {},
+        },
+      ],
     }).compile();
 
     service = module.get<ReviewsService>(ReviewsService);
