+++ src/reviews/reviews.service.ts
@@ -6,9 +6,9 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { User, Review, FavoriteReview, ReviewState } from '@prisma/client';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { PrismaService } from '../../src/prisma/prisma.service';
 import { ReviewDto } from './dto/reviews.dto';
-import { REDIS_CLIENT } from 'src/provider/redis.provider';
+import { REDIS_CLIENT } from '../../src/provider/redis.provider';
 import Redis from 'ioredis';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { RatingDto } from './dto/rating.dto';
