+++ src/user/user.e2e.spec.ts
@@ -117,320 +117,6 @@ describe('User Controller Tests', () => {
     expect(updatedUser.name).toBe('John Doe');
   });
 
-  describe('review tests', () => {
-    beforeEach(async () => {
-      await prisma.user.create({
-        data: {
-          id: '2',
-          email: 'janedoe@example.com',
-          name: 'Jane Doe',
-          isEmailVerified: true,
-          authType: AuthType.EMAIL,
-        },
-      });
-
-      await prisma.review.createMany({
-        data: [
-          {
-            postedToId: '2',
-            postedById: '1',
-            professionalism: 5,
-            reliability: 5,
-            communication: 5,
-            comment: 'Something',
-          },
-          {
-            postedToId: '2',
-            postedById: '1',
-            professionalism: 5,
-            reliability: 5,
-            communication: 5,
-          },
-        ],
-      });
-
-      await prisma.review.create({
-        data: {
-          postedToId: '1',
-          postedById: '2',
-          professionalism: 5,
-          reliability: 5,
-          communication: 5,
-        },
-      });
-    });
-
-    it('should not be able to rate a user that does not exist', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/3',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 5,
-          reliability: 5,
-          communication: 5,
-        },
-      });
-
-      expect(response.statusCode).toBe(404);
-      expect(response.json().message).toBe('User not found');
-    });
-
-    it('should throw validation error if professionalism rating is not provided', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          reliability: 5,
-          communication: 5,
-        },
-      });
-
-      expect(response.statusCode).toBe(400);
-    });
-
-    it('should throw validation error if reliability rating is not provided', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 5,
-          communication: 5,
-        },
-      });
-
-      expect(response.statusCode).toBe(400);
-    });
-
-    it('should throw validation error if communication rating is not provided', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 5,
-          reliability: 5,
-        },
-      });
-
-      expect(response.statusCode).toBe(400);
-    });
-
-    it('should not allow a user to rate themselves', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/1',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 5,
-          reliability: 5,
-          communication: 5,
-        },
-      });
-
-      expect(response.statusCode).toBe(400);
-      expect(response.json().message).toBe('You cannot rate yourself');
-    });
-
-    it('should not be able to rate anything lesser than 1', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 0,
-          reliability: 0,
-          communication: 0,
-        },
-      });
-
-      expect(response.statusCode).toBe(400);
-    });
-
-    it('should not be able to rate anything greater than 5', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 6,
-          reliability: 6,
-          communication: 6,
-        },
-      });
-
-      expect(response.statusCode).toBe(400);
-    });
-
-    it('should be able to rate another user', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 5,
-          reliability: 5,
-          communication: 5,
-        },
-      });
-
-      expect(response.statusCode).toBe(201);
-      expect(response.json().postedToId).toBe('2');
-      expect(response.json().postedById).toBe('1');
-      expect(response.json().professionalism).toBe(5);
-      expect(response.json().reliability).toBe(5);
-      expect(response.json().communication).toBe(5);
-
-      const rating = await prisma.review.findUnique({
-        where: {
-          id: response.json().id,
-        },
-      });
-      expect(rating).toBeDefined();
-      expect(rating.postedToId).toBe('2');
-      expect(rating.postedById).toBe('1');
-      expect(rating.professionalism).toBe(5);
-      expect(rating.reliability).toBe(5);
-      expect(rating.communication).toBe(5);
-    });
-
-    it('should be able to rate another user', async () => {
-      const response = await app.inject({
-        method: 'POST',
-        url: '/user/rate/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-        payload: {
-          professionalism: 5,
-          reliability: 5,
-          communication: 5,
-          anonymous: true,
-        },
-      });
-
-      expect(response.statusCode).toBe(201);
-      expect(response.json().postedToId).toBe('2');
-      expect(response.json().postedById).toBe(null);
-      expect(response.json().professionalism).toBe(5);
-      expect(response.json().reliability).toBe(5);
-      expect(response.json().communication).toBe(5);
-      expect(response.json().anonymous).toBe(true);
-
-      const rating = await prisma.review.findUnique({
-        where: {
-          id: response.json().id,
-        },
-      });
-      expect(rating).toBeDefined();
-      expect(rating.postedToId).toBe('2');
-      expect(rating.postedById).toBe(null);
-      expect(rating.professionalism).toBe(5);
-      expect(rating.reliability).toBe(5);
-      expect(rating.communication).toBe(5);
-      expect(rating.anonymous).toBe(true);
-    });
-
-    it('should be able to get the ratings of the current user', async () => {
-      const response = await app.inject({
-        method: 'GET',
-        url: '/user/ratings/self',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-      });
-
-      expect(response.statusCode).toBe(200);
-      expect(response.json()).toHaveLength(1);
-
-      const rating = response.json()[0];
-      expect(rating.userName).toBe('Jane Doe');
-      expect(rating.professionalism).toBe(5);
-      expect(rating.reliability).toBe(5);
-      expect(rating.communication).toBe(5);
-    });
-
-    it('should be able to get the ratings of another user', async () => {
-      const response = await app.inject({
-        method: 'GET',
-        url: '/user/ratings/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-      });
-
-      expect(response.statusCode).toBe(200);
-      expect(response.json()).toHaveLength(2);
-
-      const rating = response.json()[0];
-      expect(rating.userName).toBe('John Doe');
-      expect(rating.professionalism).toBe(5);
-      expect(rating.reliability).toBe(5);
-      expect(rating.communication).toBe(5);
-      expect(rating.comment).toBe('Something');
-    });
-
-    it('should be able to get average ratings of another user', async () => {
-      const response = await app.inject({
-        method: 'GET',
-        url: '/user/avg-rating/2',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-      });
-
-      expect(response.statusCode).toBe(200);
-      expect(response.json().professionalism).toBe(5);
-      expect(response.json().reliability).toBe(5);
-      expect(response.json().communication).toBe(5);
-      expect(response.json().overall).toBe(5);
-    });
-
-    it('should be able to get average rating of self', async () => {
-      const response = await app.inject({
-        method: 'GET',
-        url: '/user/avg-rating/self',
-        headers: {
-          'x-e2e-user-email': 'johndoe@example.com',
-        },
-      });
-
-      expect(response.statusCode).toBe(200);
-      expect(response.json().professionalism).toBe(5);
-      expect(response.json().reliability).toBe(5);
-      expect(response.json().communication).toBe(5);
-      expect(response.json().overall).toBe(5);
-    });
-  });
-
-  it('should be able to search for a user', async () => {
-    const response = await app.inject({
-      method: 'GET',
-      url: '/user/search/John',
-    });
-
-    expect(response.statusCode).toBe(200);
-    expect(response.json()).toHaveLength(1);
-  });
-
   afterAll(async () => {
     try {
       await prisma.user.deleteMany();
